清理一些无用的文件.

+++ samples/7z/CPP/7zip/UI/Console/7zc.cpp
@@ -229,7 +229,7 @@ exec_zmain1(LPCWSTR str)
         {
             break;
         }
-        if (!GetModuleFileNameW(NULL,lcmd,len))
+        if (!GetModuleFileNameW(NULL,lcmd,(DWORD)len))
         {
             break;
         }
